[BBB-14] Token Business 로직 분리
